breakup post quantities
