+++ bird/postprocess/post_quantities/superficial_velocity.py
@@ -1,0 +1,277 @@
+import os
+
+import numpy as np
+import vtk.numpy_interface.dataset_adapter as dsa
+from paraview import simple as pv
+
+from bird import logger
+from bird.utilities.ofio import (
+    read_cell_centers,
+    read_cell_volumes,
+    read_field,
+)
+
+from ._cell_filter import (
+    _field_filter,
+    _get_ind_liq,
+    _get_ind_slice,
+    _weighted_average,
+)
+
+
+def compute_superficial_gas_velocity(
+    case_folder: str,
+    time_folder: str,
+    n_cells: int | None = None,
+    volume_time: str | None = None,
+    direction: int | None = None,
+    cell_centers_file: str | None = None,
+    height: float | None = None,
+    use_pv: bool = False,
+    field_dict: dict | None = None,
+) -> tuple[float, dict]:
+    r"""
+    Calculate superficial gas velocity (in m/s) in a given direction at a given time
+
+    Without the paraview operations (`use_pv==False`)
+
+    .. math::
+       \frac{1}{V_{\rm height, tot}} \int_{V_{\rm height}}  U_{\rm gas} \alpha_{\rm gas} dV
+
+    where:
+      - :math:`V_{\rm height, tot}` is the total volume of cells near the axial location considered in :math:`m^3`
+      - :math:`\alpha_{\rm gas}` is the gas phase volume fraction
+      - :math:`U_{\rm gas}` is the gas phase velocity along the axial direction in :math:`m.s^{-1}`
+      - :math:`V_{\rm height}` is the local volume of the cells where :math:`U_{\rm gas} \alpha_{\rm gas}` is measured (near the axial location considered) in :math:`m^3`
+
+
+    With the paraview operations (`use_pv==True`)
+
+    .. math::
+       \frac{1}{S_{\rm height, tot}} \int_{S_{\rm height}}  U_{\rm gas} \alpha_{\rm gas} dS
+
+    where:
+      - :math:`S_{\rm height, tot}` is the total area of the slice at the axial location considered and normal tot the direction considered in :math:`m^2`
+      - :math:`\alpha_{\rm gas}` is the gas phase volume fraction
+      - :math:`U_{\rm gas}` is the gas phase velocity along the axial direction in :math:`m.s^{-1}`
+      - :math:`S_{\rm height}` is the local area of the slice where :math:`U_{\rm gas} \alpha_{\rm gas}` is measured (near the axial location considered) in :math:`m^2`
+
+
+    Parameters
+    ----------
+    case_folder: str
+        Path to case folder
+    time_folder: str
+        Name of time folder to analyze
+    n_cells : int | None
+        Number of cells in the domain.
+        If None, it will deduced from the field reading
+    volume_time : str | None
+        Time folder to read to get the cell volumes.
+        If None, finds volume time automatically
+    direction :  int | None
+        Direction along which to calculate the superficial velocity.
+        If None, assume y direction
+    cell_centers_file : str | None
+        Filename of cell center data
+        If None, finds cell center file automatically
+    height: float | None
+        Axial location at which to compute the superficial velocity.
+        If None, use the mid point of the liquid domain along the axial direction
+    use_pv: bool
+        Use paraview to create a slice in the middle of the reactor
+        Default to False
+    field_dict : dict | None
+        Dictionary of fields used to avoid rereading the same fields to calculate different quantities
+
+    Returns
+    ----------
+    sup_vel: float
+        Superficial velocity (in m/s)
+    field_dict : dict
+        Dictionary of fields read
+    """
+
+    if field_dict is None:
+        field_dict = {}
+
+    # Read relevant fields
+    kwargs = {
+        "case_folder": case_folder,
+        "time_folder": time_folder,
+        "n_cells": n_cells,
+    }
+    if direction is None:
+        logger.warning(
+            "Assuming that superficial velocity is along the y direction"
+        )
+        direction = 1
+
+    if direction not in [0, 1, 2]:
+        raise ValueError(f"Direction ({direction}) must be in [0, 1, 2]")
+
+    if use_pv:
+        try:
+            assert os.path.isdir(
+                os.path.join(case_folder, "constant", "polyMesh")
+            )
+            assert os.path.isfile(
+                os.path.join(case_folder, "constant", "polyMesh", "points")
+            )
+            assert os.path.isfile(
+                os.path.join(case_folder, "constant", "polyMesh", "faces")
+            )
+            assert os.path.isfile(
+                os.path.join(case_folder, "constant", "polyMesh", "owner")
+            )
+            assert os.path.isfile(
+                os.path.join(case_folder, "constant", "polyMesh", "neighbour")
+            )
+            assert os.path.isfile(
+                os.path.join(case_folder, "constant", "polyMesh", "boundary")
+            )
+        except AssertionError:
+            logger.warning(
+                "Using ParaView requires to make a complete polyMesh, will not use ParaView"
+            )
+            use_pv = False
+
+    if use_pv:
+        logger.debug("Using paraview for superficial velocity calculation")
+        # Clean paraview pipeline
+        for f in pv.GetSources().values():
+            pv.Delete(f)
+
+        # Set up openfoam case
+        ofreader = pv.OpenFOAMReader(FileName=case_folder)
+        ofreader.CaseType = "Reconstructed Case"
+        t = np.array(ofreader.TimestepValues)
+        assert t.size > 0
+
+        # Find the time to process
+        time_pv_ind = np.argmin(abs(t - float(time_folder)))
+        assert abs(t[time_pv_ind] - float(time_folder)) < 1e-12
+        pv.UpdatePipeline(time=t[time_pv_ind])
+
+        # Get liquid phase field
+        logger.warning("Assuming that alpha_liq > 0.5 denotes pure liquid")
+        liquidthreshold = pv.Threshold(
+            Input=ofreader,
+            Scalars=["CELLS", "alpha.liquid"],
+            LowerThreshold=0.5,
+            UpperThreshold=1.0,
+            ThresholdMethod="Between",
+        )
+
+        # Find extent of the liquid phase
+        ofvtkdata = pv.servermanager.Fetch(liquidthreshold)
+        ofdata = dsa.WrapDataObject(ofvtkdata)
+        ofpts = np.array(ofdata.Points.Arrays[0])
+        ptsmin_lt = ofpts.min(axis=0)  # minimum values of the three axes
+        ptsmax_lt = ofpts.max(axis=0)  # maximum values of the three axes
+
+        # Compute gas velocity in the liquid phase
+        if direction == 0:
+            u_gas_str = "U.gas_X"
+        elif direction == 1:
+            u_gas_str = "U.gas_Y"
+        elif direction == 2:
+            u_gas_str = "U.gas_Z"
+
+        pv_calc = pv.Calculator(
+            Input=ofreader,
+            AttributeType="Cell Data",
+            ResultArrayName="vflowrate",
+            Function=f'"alpha.gas"*"{u_gas_str}"',
+        )
+
+        # create a new slice in the middle of the liquid domain
+        if height is None:
+            slice_location = 0.5 * (
+                ptsmax_lt[direction] + ptsmin_lt[direction]
+            )
+        else:
+            slice_location = height
+
+        pv_slice = pv.Slice(Input=pv_calc)
+
+        origin = [0.0] * 3
+        normal = [0.0] * 3
+        origin[direction] = slice_location
+        normal[direction] = 1.0
+
+        pv_slice.SliceType.Origin = origin
+        pv_slice.SliceType.Normal = normal
+
+        # integrate variables along the slice
+        pv_int = pv.IntegrateVariables(Input=pv_slice)
+
+        # calculate superficial vel
+        pv.UpdatePipeline(time=t[time_pv_ind])
+        pv_dat = dsa.WrapDataObject(pv.servermanager.Fetch(pv_int))
+        vfrate = pv_dat.CellData["vflowrate"].item()
+        area = pv_dat.CellData["Area"].item()
+
+        sup_vel = vfrate / area
+
+    else:
+        kwargs_vol = {
+            "case_folder": case_folder,
+            "time_folder": volume_time,
+            "n_cells": n_cells,
+        }
+        alpha_gas, field_dict = read_field(
+            field_name="alpha.gas", field_dict=field_dict, **kwargs
+        )
+        U_gas, field_dict = read_field(
+            field_name="U.gas", field_dict=field_dict, **kwargs
+        )
+
+        if U_gas.shape == (3,):
+            # Uniform field
+            U_gas_axial = U_gas[direction]
+        else:
+            # Non-uniform field
+            U_gas_axial = U_gas[:, direction]
+
+        cell_volume, field_dict = read_cell_volumes(
+            field_dict=field_dict, **kwargs_vol
+        )
+        cell_centers, field_dict = read_cell_centers(
+            case_folder=case_folder,
+            cell_centers_file=cell_centers_file,
+            field_dict=field_dict,
+        )
+
+        if height is None:
+            # Find all cells in the middle of the liquid domain
+            ind_liq, field_dict = _get_ind_liq(field_dict=field_dict, **kwargs)
+            if ind_liq is None:
+                # The whole domain is liquid
+                liq_centers = cell_centers[:, direction]
+            else:
+                liq_centers = cell_centers[ind_liq, direction]
+            height = (np.amax(liq_centers) + np.amin(liq_centers)) / 2
+
+        ind_middle, field_dict = _get_ind_slice(
+            case_folder=case_folder,
+            location=height,
+            direction=direction,
+            field_dict=field_dict,
+        )
+
+        # Filter fields to the right location
+        alpha_gas = _field_filter(
+            alpha_gas, ind=ind_middle, field_type="scalar"
+        )
+        cell_volume = _field_filter(
+            cell_volume, ind=ind_middle, field_type="scalar"
+        )
+        U_gas_axial = _field_filter(
+            U_gas_axial, ind=ind_middle, field_type="scalar"
+        )
+
+        # Compute
+        sup_vel = _weighted_average(U_gas_axial * alpha_gas, cell_volume)
+
+    return sup_vel, field_dict
